Initial application creation request
I want to make a single-page web page. The goal is to be able to track
how long until certain times will occur.

1. Users should be able to add new times along with descriptions. For
example, I might want to have a countdown to next Christmas as well as
an arbitrary date like November 2nd, 2029.

2. The dates and descriptions should be persisted in the URL through
some sort of encoding (I don't care how), so that I can send a link
with the encoded dates to share with friends.

3. No server dependencies, everything should be in a single page with
HTML and inline CSS and HTML.