Add ability to stop past moments
Can you make it so that dates being tracked in the past can be stopped?
For example, if someone was tracking how long it has been since they've
been to the zoo and then they go to the zoo, they would stop the
tracker, leaving it listed but frozen in time.